Extract inlined sql scalar functions into a plugin module
