+++ pom.xml
@@ -216,6 +216,7 @@
         <module>presto-function-server</module>
         <module>presto-router-example-plugin-scheduler</module>
         <module>presto-plan-checker-router-plugin</module>
+        <module>presto-sql-invoked-functions-plugin</module>
     </modules>
 
     <dependencyManagement>
