+++ presto-main-base/src/main/java/com/facebook/presto/operator/scalar/ArrayIntersectFunction.java
@@ -19,8 +19,6 @@ import com.facebook.presto.operator.aggregation.OptimizedTypedSet;
 import com.facebook.presto.spi.function.Description;
 import com.facebook.presto.spi.function.OperatorDependency;
 import com.facebook.presto.spi.function.ScalarFunction;
-import com.facebook.presto.spi.function.SqlInvokedScalarFunction;
-import com.facebook.presto.spi.function.SqlParameter;
 import com.facebook.presto.spi.function.SqlType;
 import com.facebook.presto.spi.function.TypeParameter;
 
@@ -59,15 +57,5 @@ public final class ArrayIntersectFunction
         typedSet.intersect(leftArray);
 
         return typedSet.getBlock();
-    }
-
-    @SqlInvokedScalarFunction(value = "array_intersect", deterministic = true, calledOnNullInput = false)
-    @Description("Intersects elements of all arrays in the given array")
-    @TypeParameter("T")
-    @SqlParameter(name = "input", type = "array<array<T>>")
-    @SqlType("array<T>")
-    public static String arrayIntersectArray()
-    {
-        return "RETURN reduce(input, IF((cardinality(input) = 0), ARRAY[], input[1]), (s, x) -> array_intersect(s, x), (s) -> s)";
     }
 }
