+++ presto-main/etc/config.properties
@@ -50,8 +50,7 @@ plugin.bundles=\
   ../presto-cluster-ttl-providers/pom.xml,\
   ../presto-node-ttl-fetchers/pom.xml,\
   ../presto-hive-function-namespace/pom.xml,\
-  ../presto-delta/pom.xml,\
-  ../presto-hudi/pom.xml
+  ../presto-sql-invoked-functions-plugin/pom.xml
 
 presto.version=testversion
 node-scheduler.include-coordinator=true
