+++ presto-sql-invoked-functions-plugin/src/main/java/com/facebook/presto/scalar/sql/SimpleSamplingPercent.java
@@ -11,25 +11,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.operator.scalar.sql;
+package com.facebook.presto.scalar.sql;
 
 import com.facebook.presto.spi.function.Description;
 import com.facebook.presto.spi.function.SqlInvokedScalarFunction;
 import com.facebook.presto.spi.function.SqlParameter;
 import com.facebook.presto.spi.function.SqlType;
 
-@SqlInvokedScalarFunction(value = "map_normalize", deterministic = true, calledOnNullInput = false)
-@Description("Returns the map with the same keys but all non-null values are scaled proportionally so that the sum of values becomes 1.")
-public class MapNormalizeFunction
+public class SimpleSamplingPercent
 {
-    private MapNormalizeFunction() {}
+    private SimpleSamplingPercent() {}
 
-    @SqlParameter(name = "input", type = "map<varchar, double>")
-    @SqlType("map<varchar, double>")
-    public static String arraySumDouble()
+    @SqlInvokedScalarFunction(value = "key_sampling_percent", deterministic = true, calledOnNullInput = false)
+    @Description("Returns a value between 0.0 and 1.0 using the hash of the given input string")
+    @SqlParameter(name = "input", type = "varchar")
+    @SqlType("double")
+    public static String keySamplingPercent()
     {
-        return "RETURN " +
-                "transform(array[ROW(input, cast(array_sum(map_values(input)) as double))], " +
-                "x->transform_values(x[1], (k, v) -> (v / x[2])))[1]";
+        return "return (abs(from_ieee754_64(xxhash64(cast(input as varbinary)))) % 100) / 100. ";
     }
 }
